Make activities persistent
The android lowmemorykiller kernel driver is killing processes based on their
oom_adj value which is managed by the AcitivityManagerService. The higher the
value the more likely it is that the lowmemorykiller will kill the process.
This patch adds support to define a max value for an activity to keep them
running. This is usefull to speed up loading core activities like the phone or
contacts.